+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Only partial changelog, [commit history](https://framagit.org/SondagePro-LimeSurvey-plugin/sendMailCron/commits/master) show all changelog.
 
+## Not released [0.3.2]
+
+### Fix
+
+- Send after reminder tested
+
+
 ## [0.3.1] - 2017-03-16
 
 ### Feature